chore: code clean up
